+++ doc/changes.rst
@@ -22,6 +22,7 @@ Bug fixes
 * `#497 <https://bitbucket.org/openpyxl/openpyxl/issues/497/>`_ Scatter charts are broken
 * `#499 <https://bitbucket.org/openpyxl/openpyxl/issues/499/>`_ Inconsistent conversion of localised datetimes
 * `#500 <https://bitbucket.org/openpyxl/openpyxl/issues/500/>`_ Adding images leads to unreadable files
+* `#509 <https://bitbucket.org/openpyxl/openpyxl/issues/509/>`_ Improve handling of sheet names
 
 
 Minor changes